refactor: replace old layerchart cashflow with grid-based implementation
