+++ src/routes/(app)/big-picture/+page.svelte
@@ -9,7 +9,6 @@
 	import { m } from '$lib/paraglide/messages';
 	import { getPocketBaseContext } from '$lib/pocketbase.svelte';
 
-	import CashflowGrid from './cashflow-grid.svelte';
 	import Cashflow from './cashflow.svelte';
 	import Summary from './summary.svelte';
 	import TrailingCashflow from './trailing-cashflow.svelte';
@@ -36,10 +35,6 @@
 	<Section>
 		<SectionTitle title="Summary" />
 		<Summary />
-	</Section>
-
-	<Section>
-		<CashflowGrid />
 	</Section>
 
 	<Section>
